modularize outbound
